pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "penit"
version = "2.0.0"
description = "Full-screen overlay drawing application"
requires-python = ">=3.10"
dependencies = ["PyQt6>=6.5"]

[project.optional-dependencies]
x11 = ["python-xlib>=0.33"]
wayland = ["dbus-python>=1.3"]
macos = ["pyobjc-framework-Cocoa>=9.0", "pyobjc-framework-Quartz>=9.0"]

[project.scripts]
penit = "penit.__main__:main"

[tool.setuptools.packages.find]
include = ["penit*"]
